// MsfViewPCF/cds-generated/entities/axa_SalesFulfillmentStatus.ts
export const axa_salesfulfillmentstatusMetadata = {
  typeName: "mscrm.axa_salesfulfillmentstatus",
  logicalName: "axa_salesfulfillmentstatus",
  collectionName: "axa_salesfulfillmentstatuses",
  primaryIdAttribute: "axa_salesfulfillmentstatusid",
  attributeTypes: {
    // Numeric Types
    importsequencenumber: "Integer",
    timezoneruleversionnumber: "Integer",
    utcconversiontimezonecode: "Integer",
    versionnumber: "BigInt",
    // Optionsets
    axa_currentmilestone: "Optionset",
    axa_currentphase: "Optionset",
    axa_doescustomerhavecws: "Optionset",
    axa_doescustomerhavedatagovernanceform: "Optionset",
    axa_locationbranch: "Optionset",
    axa_salesstatus: "Optionset",
    axa_typeofsale1: "Optionset",
    statecode: "Optionset",
    statuscode: "Optionset",
    // Date Formats
    axa_confirmeddeliverydate: "DateOnly:UserLocal",
    axa_esd: "DateOnly:UserLocal",
    axa_estimatedcustomerreceiptdate: "DateOnly:UserLocal",
    axa_estimateddeliverytocustomer: "DateOnly:UserLocal",
    axa_estimatedpartsorderdate: "DateOnly:UserLocal",
    axa_estimatedpartsorderdate1: "DateOnly:UserLocal",
    axa_machinedeliveredtocustomer: "DateOnly:UserLocal",
    axa_projectedservicestartdate: "DateOnly:UserLocal",
    axa_projectedservicestartdate1: "DateOnly:UserLocal",
    createdon: "DateAndTime:UserLocal",
    modifiedon: "DateAndTime:UserLocal",
    overriddencreatedon: "DateOnly:UserLocal",
  },
  navigation: {
    owninguser: ["mscrm.systemuser"],
    owningteam: ["mscrm.team"],
    owningbusinessunit: ["mscrm.businessunit"],
    ownerid: ["mscrm.principal"],
    modifiedonbehalfby: ["mscrm.systemuser"],
    modifiedby: ["mscrm.systemuser"],
    createdonbehalfby: ["mscrm.systemuser"],
    createdby: ["mscrm.systemuser"],
    axa_ServiceCallnumber: ["mscrm.incident"],
    axa_SalesResponsible: ["mscrm.systemuser"],
    axa_Mocel: ["mscrm.z2t_model"],
    axa_Make: ["mscrm.z2t_make"],
    axa_EquipmentNumberEQN: ["mscrm.z2t_equipment"],
    axa_DSF: ["mscrm.axa_dealsetupform"],
    axa_CustomerProspect: ["mscrm.account"],
  },
};

// Attribute constants
export enum axa_SalesFulfillmentStatusAttributes {
  axa_ConfirmedDeliveryDate = "axa_confirmeddeliverydate",
  axa_CurrentMilestone = "axa_currentmilestone",
  axa_CurrentPhase = "axa_currentphase",
  axa_CustomerName = "axa_customername",
  axa_CustomerPO = "axa_customerpo",
  axa_CustomerProspect = "axa_customerprospect",
  axa_CustomerProspectName = "axa_customerprospectname",
  axa_CustomerProspectYomiName = "axa_customerprospectyominame",
  axa_CWSID = "axa_cwsid",
  axa_Description = "axa_description",
  axa_DoescustomerhaveCWS = "axa_doescustomerhavecws",
  axa_DoesCustomerhavedatagovernanceform = "axa_doescustomerhavedatagovernanceform",
  axa_DSF = "axa_dsf",
  axa_DSFName = "axa_dsfname",
  axa_EquipmentNumberEQN = "axa_equipmentnumbereqn",
  axa_EquipmentNumberEQNName = "axa_equipmentnumbereqnname",
  axa_ESD = "axa_esd",
  axa_EstimatedCustomerReceiptDate = "axa_estimatedcustomerreceiptdate",
  axa_EstimatedDeliverytoCustomer = "axa_estimateddeliverytocustomer",
  axa_EstimatedPartsOrderDate = "axa_estimatedpartsorderdate",
  axa_EstimatedPartsOrderDate1 = "axa_estimatedpartsorderdate1",
  axa_Instock = "axa_instock",
  axa_Internalfinancingrequired = "axa_internalfinancingrequired",
  axa_LocationBranch = "axa_locationbranch",
  axa_MachineDeliveredtoCustomer = "axa_machinedeliveredtocustomer",
  axa_Machinepurchaseordernumber = "axa_machinepurchaseordernumber",
  axa_Machinesalesordernumber = "axa_machinesalesordernumber",
  axa_Make = "axa_make",
  axa_MakeName = "axa_makename",
  axa_Mocel = "axa_mocel",
  axa_MocelName = "axa_mocelname",
  axa_Name = "axa_name",
  axa_ProjectedServiceStartDate = "axa_projectedservicestartdate",
  axa_ProjectedServiceStartDate1 = "axa_projectedservicestartdate1",
  axa_SalesFulfillmentStatusId = "axa_salesfulfillmentstatusid",
  axa_SalesResponsible = "axa_salesresponsible",
  axa_SalesResponsibleName = "axa_salesresponsiblename",
  axa_SalesResponsibleYomiName = "axa_salesresponsibleyominame",
  axa_SalesStatus = "axa_salesstatus",
  axa_Serialnumber = "axa_serialnumber",
  axa_ServiceCallnumber = "axa_servicecallnumber",
  axa_ServiceCallnumberName = "axa_servicecallnumbername",
  axa_Statustxt = "axa_statustxt",
  axa_Tradeinincluded = "axa_tradeinincluded",
  axa_TypeofSale = "axa_typeofsale",
  axa_TypeofSale1 = "axa_typeofsale1",
  axa_Warehouse = "axa_warehouse",
  CreatedBy = "createdby",
  CreatedByName = "createdbyname",
  CreatedByYomiName = "createdbyyominame",
  CreatedOn = "createdon",
  CreatedOnBehalfBy = "createdonbehalfby",
  CreatedOnBehalfByName = "createdonbehalfbyname",
  CreatedOnBehalfByYomiName = "createdonbehalfbyyominame",
  ImportSequenceNumber = "importsequencenumber",
  ModifiedBy = "modifiedby",
  ModifiedByName = "modifiedbyname",
  ModifiedByYomiName = "modifiedbyyominame",
  ModifiedOn = "modifiedon",
  ModifiedOnBehalfBy = "modifiedonbehalfby",
  ModifiedOnBehalfByName = "modifiedonbehalfbyname",
  ModifiedOnBehalfByYomiName = "modifiedonbehalfbyyominame",
  OverriddenCreatedOn = "overriddencreatedon",
  OwnerId = "ownerid",
  OwnerIdName = "owneridname",
  OwnerIdType = "owneridtype",
  OwnerIdYomiName = "owneridyominame",
  OwningBusinessUnit = "owningbusinessunit",
  OwningBusinessUnitName = "owningbusinessunitname",
  OwningTeam = "owningteam",
  OwningUser = "owninguser",
  statecode = "statecode",
  statuscode = "statuscode",
  TimeZoneRuleVersionNumber = "timezoneruleversionnumber",
  UTCConversionTimeZoneCode = "utcconversiontimezonecode",
  VersionNumber = "versionnumber",
}